Perubahan tipe data request dari MasGusti

+++ database/migrations/2021_07_01_045229_create_transactions_table.php
@@ -19,12 +19,12 @@ class CreateTransactionsTable extends Migration
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
             $table->string('transaction_code');
             $table->date('transaction_date');
-            $table->integer('total_item');
-            $table->float('total_price', 16,2);
-            $table->float('discount', 16,2)->nullable();
-            $table->float('price_after_discount', 16,2);
-            $table->float('pajak_ppn', 16,2)->nullable();
-            $table->float('sub_total_price', 16,2);
+            $table->bigInteger('total_item');
+            $table->bigInteger('total_price');
+            $table->bigInteger('discount')->nullable();
+            $table->bigInteger('price_after_discount');
+            $table->bigInteger('pajak_ppn')->nullable();
+            $table->bigInteger('sub_total_price');
             $table->integer('status')->default(0); // complete payment: 1, not-complete: 0, shipping: 2
             $table->timestamps();
         });
